backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pydantic
sqlalchemy
psycopg2-binary
easyocr
opencv-python
pillow
python-multipart